move RequestDispatcher class to internal.http package
