+++ components/org.wso2.carbon.uis/src/main/java/org/wso2/carbon/uis/internal/http/RequestDispatcher.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.uis.internal;
+package org.wso2.carbon.uis.internal.http;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,8 +24,6 @@ import org.wso2.carbon.uis.api.App;
 import org.wso2.carbon.uis.api.exception.UISRuntimeException;
 import org.wso2.carbon.uis.api.http.HttpRequest;
 import org.wso2.carbon.uis.api.http.HttpResponse;
-import org.wso2.carbon.uis.internal.http.PageRequestDispatcher;
-import org.wso2.carbon.uis.internal.http.ResponseBuilder;
 import org.wso2.carbon.uis.internal.io.StaticRequestDispatcher;
 
 /**
