+++ components/org.wso2.carbon.uis/src/test/java/org/wso2/carbon/uis/internal/http/RequestDispatcherTest.java
@@ -16,14 +16,13 @@
  * under the License.
  */
 
-package org.wso2.carbon.uis.internal;
+package org.wso2.carbon.uis.internal.http;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.wso2.carbon.uis.api.exception.UISRuntimeException;
 import org.wso2.carbon.uis.api.http.HttpRequest;
 import org.wso2.carbon.uis.api.http.HttpResponse;
-import org.wso2.carbon.uis.internal.http.PageRequestDispatcher;
 import org.wso2.carbon.uis.internal.io.StaticRequestDispatcher;
 
 import static org.mockito.ArgumentMatchers.any;
